Moved node class into its own header file.
